+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/internal/storage/db/H2DatabaseConfiguration.java
@@ -55,5 +55,10 @@ public class H2DatabaseConfiguration extends DatabaseConfiguration {
     public String castToEnumType(String variable, String type) {
       return variable;
     }
+
+    @Override
+    public boolean hasUpdateableCTE() {
+      return false;
+    }
   }
 }