+++ nflow-engine/src/main/java/com/nitorcreations/nflow/engine/internal/storage/db/PgDatabaseConfiguration.java
@@ -32,5 +32,10 @@ public class PgDatabaseConfiguration extends DatabaseConfiguration {
     public String castToEnumType(String variable, String type) {
       return variable + "::" + type;
     }
+
+    @Override
+    public boolean hasUpdateableCTE() {
+      return true;
+    }
   }
 }